requirements.txt
panda3d==1.10.8
Pillow==8.1.0
PyYAML==5.3.1
